backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.2
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.4.0
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
